+++ orc8r/cloud/go/services/directoryd/servicers/protected/id_gen_test.go
@@ -22,7 +22,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
-	"magma/orc8r/cloud/go/services/directoryd/servicers"
+	servicers "magma/orc8r/cloud/go/services/directoryd/servicers/protected"
 	magmaerrors "magma/orc8r/lib/go/errors"
 )
