+++ orc8r/cloud/go/services/directoryd/servicers/protected/servicer_test.go
@@ -20,7 +20,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"magma/orc8r/cloud/go/blobstore"
-	"magma/orc8r/cloud/go/services/directoryd/servicers"
+	servicers "magma/orc8r/cloud/go/services/directoryd/servicers/protected"
 	"magma/orc8r/cloud/go/services/directoryd/storage"
 	stateTestInit "magma/orc8r/cloud/go/services/state/test_init"
 	"magma/orc8r/cloud/go/sqorc"
